In Account, I created a code refactoring for the DAO and a test for the methods.
