+++ src/main/java/pl/wasik/damian/java/app/bank/exception/account/ReadAccountException.java
@@ -1,4 +1,4 @@
-package pl.wasik.damian.java.app.bank.exception.read;
+package pl.wasik.damian.java.app.bank.exception.account;
 
 import pl.wasik.damian.java.app.bank.exception.AccountException;
