+++ src/main/java/pl/wasik/damian/java/app/bank/exception/account/UpdateAccountException.java
@@ -1,4 +1,4 @@
-package pl.wasik.damian.java.app.bank.exception.update;
+package pl.wasik.damian.java.app.bank.exception.account;
 
 import pl.wasik.damian.java.app.bank.exception.AccountException;
